rename file + simplify isVueScriptTag usage
